requirements.txt
fastapi~=0.121.3
uvicorn~=0.38.0
batata-lib~=0.1.6